refactor(domain): split jvmMain repo interfaces into commonMain base + JvmXxx extension
Each repository that previously lived wholly in jvmMain (because it
returned java.io.File from some methods) is now split into two layers:
- commonMain base interface: File-free methods (download, remove,
  isDownloaded, isCached, saveReadingPosition, etc.)
- jvmMain JvmXxx extension interface: File-returning methods (openEpub,
  openPdf, openCbz, openEpubForMetadata, bundleFile, readTrack, etc.)

Repositories split: EpubRepository, PdfRepository, CbzRepository,
ReadaloudAudioRepository, AudiobookDownloadRepository,
AudiobookCacheRepository.

Concrete implementations now implement the JvmXxx interfaces; Koin
binds both levels so existing get<BaseType>() call sites continue to
work alongside new get<JvmXxx>() consumers.

This is a prerequisite for porting LibraryItemDetailViewModel to
feature:library/commonMain (issue #952).
