+++ core/data/src/androidMain/kotlin/com/riffle/core/data/StorytellerBundleAudiobookSource.kt
@@ -2,7 +2,7 @@ package com.riffle.core.data
 
 import com.riffle.core.domain.AudiobookSession
 import com.riffle.core.domain.BundleAudiobookSource
-import com.riffle.core.domain.ReadaloudAudioRepository
+import com.riffle.core.domain.JvmReadaloudAudioRepository
 import com.riffle.core.models.ReadaloudLink
 import com.riffle.core.domain.ReadaloudLinkRepository
 
@@ -18,7 +18,7 @@ import com.riffle.core.domain.ReadaloudLinkRepository
  */
 class StorytellerBundleAudiobookSource(
     private val readaloudLinkRepository: ReadaloudLinkRepository,
-    private val readaloudAudioRepository: ReadaloudAudioRepository,
+    private val readaloudAudioRepository: JvmReadaloudAudioRepository,
     private val linksByAbsItem: OfflineAvailabilitySnapshot<String, ReadaloudLink>,
 ) : BundleAudiobookSource {
