+++ core/domain/src/commonMain/kotlin/com/riffle/core/domain/AudiobookCacheRepository.kt
@@ -1,0 +1,6 @@
+package com.riffle.core.domain
+
+interface AudiobookCacheRepository {
+    fun isCached(sourceId: String, itemId: String): Boolean
+    suspend fun remove(sourceId: String, itemId: String): Long
+}
