+++ core/domain/src/commonMain/kotlin/com/riffle/core/domain/CbzRepository.kt
@@ -69,3 +69,8 @@ interface CbzRepository {
      */
     suspend fun awaitCachedSource(item: LibraryItem): CbzLocalSource?
 }
+
+/** JVM extension of [CbzRepository] adding [java.io.File]-returning methods for Android/JVM hosts. */
+interface JvmCbzRepository : CbzRepository {
+    suspend fun awaitCachedFile(item: LibraryItem): java.io.File?
+}
