+++ core/domain/src/jvmMain/kotlin/com/riffle/core/domain/AudiobookDownloadRepository.kt
@@ -1,30 +1,10 @@
 package com.riffle.core.domain
 
-sealed class AudiobookDownloadResult {
-    data object Success : AudiobookDownloadResult()
-    data class NetworkError(val cause: Throwable) : AudiobookDownloadResult()
-}
-
 /**
- * Permanent offline copy of an [Audiobook] — the ebook-Download analogue for audio (ADR 0035). An
- * audiobook is several ABS tracks, so a download is a *directory* of track files plus a manifest that
- * records the timeline (per-track offsets/durations + chapters) so playback reconstructs the book
- * offline without re-opening an ABS play session. For the auto-cache (evictable) tier see
- * [AudiobookCacheRepository].
+ * JVM extension of [AudiobookDownloadRepository] that exposes [localSession] — backed by
+ * [AudiobookSession] which carries a [java.io.File] ref and therefore cannot be in commonMain.
  */
-interface AudiobookDownloadRepository {
-    fun isDownloaded(sourceId: String, itemId: String): Boolean
-
+interface JvmAudiobookDownloadRepository : AudiobookDownloadRepository {
     /** A playable session backed by the downloaded local files (`file://` track URLs), or null. */
     fun localSession(sourceId: String, itemId: String): AudiobookSession?
-
-    /** Downloads every track of the audiobook to permanent storage, reporting cumulative bytes. */
-    suspend fun download(
-        sourceId: String,
-        itemId: String,
-        onProgress: (downloaded: Long, total: Long) -> Unit,
-    ): AudiobookDownloadResult
-
-    /** Removes the downloaded copy; returns bytes freed. */
-    suspend fun remove(sourceId: String, itemId: String): Long
 }
