+++ core/domain/src/jvmMain/kotlin/com/riffle/core/domain/PdfRepository.kt
@@ -13,21 +13,7 @@ sealed class PdfOpenResult {
     data object Offline : PdfOpenResult()
 }
 
-sealed class PdfDownloadResult {
-    data object Success : PdfDownloadResult()
-    data object AlreadyDownloaded : PdfDownloadResult()
-    data class NetworkError(val cause: Throwable) : PdfDownloadResult()
-}
-
-interface PdfRepository {
+interface JvmPdfRepository : PdfRepository {
     suspend fun openPdf(item: LibraryItem): PdfOpenResult
     suspend fun openPdfForMetadata(item: LibraryItem): PdfOpenResult = openPdf(item)
-    suspend fun downloadPdf(
-        item: LibraryItem,
-        onProgress: (downloaded: Long, total: Long) -> Unit = { _, _ -> },
-    ): PdfDownloadResult
-    suspend fun removeDownload(sourceId: String, itemId: String)
-    fun isDownloaded(sourceId: String, itemId: String): Boolean
-    fun isCached(sourceId: String, itemId: String): Boolean
-    suspend fun saveReadingPosition(itemId: String, locatorJson: String)
 }
